examples/requirements/testing.txt
#
# This file is autogenerated by pip-compile with python 3.6
# To update, run:
#
#    pip-compile testing.in
#
asgiref==3.4.1
    # via django
astroid==2.9.0
    # via pylint
attrs==21.2.0
    # via
    #   jsonschema
    #   pytest
backports.entry-points-selectable==1.1.1
    # via virtualenv
black==21.11b1
    # via -r style_checkers.in
click==8.0.3
    # via black
coverage[toml]==6.2
    # via
    #   -r test.in
    #   pytest-cov
dataclasses==0.8
    # via black
distlib==0.3.3
    # via virtualenv
django==3.2.9
    # via
    #   -r django_3_2.in
    #   django-debug-toolbar
    #   django-nine
    #   djangorestframework
    #   drf-spectacular
    #   drf-spectacular-sidecar
django-debug-toolbar==3.2.2
    # via -r django_3_2.in
django-debug-toolbar-force==0.1.8
    # via -r django_3_2.in
django-nine==0.2.5
    # via
    #   django-debug-toolbar-force
    #   django-ormex
django-ormex==0.2.1
    # via -r django_3_2.in
djangorestframework==3.12.4
    # via
    #   -r django_3_2.in
    #   drf-spectacular
drf-spectacular[sidecar]==0.21.0
    # via -r django_3_2.in
drf-spectacular-sidecar==2021.12.6
    # via drf-spectacular
factory-boy==3.2.1
    # via -r test.in
faker==9.9.0
    # via
    #   -r test.in
    #   factory-boy
filelock==3.4.0
    # via
    #   tox
    #   virtualenv
importlib-metadata==4.8.2
    # via
    #   backports.entry-points-selectable
    #   click
    #   jsonschema
    #   pluggy
    #   pytest
    #   tox
    #   virtualenv
importlib-resources==5.4.0
    # via virtualenv
inflection==0.5.1
    # via drf-spectacular
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via pylint
jsonschema==3.2.0
    # via drf-spectacular
lazy-object-proxy==1.6.0
    # via astroid
mccabe==0.6.1
    # via pylint
mypy-extensions==0.4.3
    # via black
packaging==21.3
    # via
    #   pytest
    #   tox
pathspec==0.9.0
    # via black
pillow==8.4.0
    # via -r common.in
platformdirs==2.4.0
    # via
    #   black
    #   pylint
    #   virtualenv
pluggy==0.13.1
    # via
    #   pytest
    #   tox
py==1.11.0
    # via
    #   -r test.in
    #   pytest
    #   tox
pycodestyle==2.8.0
    # via -r style_checkers.in
pylint==2.12.1
    # via -r style_checkers.in
pyparsing==3.0.6
    # via packaging
pyrsistent==0.18.0
    # via jsonschema
pytest==6.2.4
    # via
    #   -r test.in
    #   pytest-cov
    #   pytest-django
    #   pytest-ordering
    #   pytest-pythonpath
pytest-cov==2.12.0
    # via -r test.in
pytest-django==4.3.0
    # via -r test.in
pytest-ordering==0.6
    # via -r test.in
pytest-pythonpath==0.7.3
    # via -r test.in
python-dateutil==2.8.2
    # via faker
python-memcached==1.58
    # via -r common.in
pytz==2021.3
    # via
    #   -r common.in
    #   django
pyyaml==6.0
    # via drf-spectacular
regex==2021.11.10
    # via black
six==1.16.0
    # via
    #   -r common.in
    #   django-debug-toolbar-force
    #   django-ormex
    #   jsonschema
    #   python-dateutil
    #   python-memcached
    #   tox
    #   virtualenv
sqlparse==0.4.2
    # via
    #   django
    #   django-debug-toolbar
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via
    #   pylint
    #   pytest
    #   tox
tomli==1.2.2
    # via
    #   black
    #   coverage
tox==3.23.1
    # via -r test.in
typed-ast==1.5.1
    # via
    #   astroid
    #   black
typing-extensions==4.0.1
    # via
    #   asgiref
    #   astroid
    #   black
    #   drf-spectacular
    #   faker
    #   importlib-metadata
    #   pylint
uritemplate==4.1.1
    # via drf-spectacular
virtualenv==20.10.0
    # via tox
wrapt==1.13.3
    # via astroid
zipp==3.6.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
